🔢 The Matrix: there is no float64, only int32
BackoffMultiplier transcends float64 illusion - percentage FTW
- CRDs now use int32 (200 = 2.0x, 150 = 1.5x)
- Conversion layer divides by 100 for internal float64
- Controller extracts int64 from unstructured, casts to int32
- allowDangerousTypes removed from Makefile - we're off the grid
- Cross-language compat achieved, warnings eliminated

+++ pkg/apis/uds/v1alpha2/types.go
@@ -245,12 +245,12 @@ type RetryPolicy struct {
 	// +kubebuilder:default="5m"
 	MaxBackoff string `json:"maxBackoff,omitempty"`
 
-	// BackoffMultiplier for exponential backoff
+	// BackoffMultiplier for exponential backoff as a percentage (200 = 2.0x, 150 = 1.5x)
 	// +optional
-	// +kubebuilder:default=2.0
-	// +kubebuilder:validation:Type=number
-	// +kubebuilder:validation:crd:allowDangerousTypes=true
-	BackoffMultiplier *float64 `json:"backoffMultiplier,omitempty"`
+	// +kubebuilder:default=200
+	// +kubebuilder:validation:Minimum=100
+	// +kubebuilder:validation:Maximum=1000
+	BackoffMultiplier *int32 `json:"backoffMultiplier,omitempty"`
 
 	// RetryableErrors defines which error patterns should trigger retries
 	// Supports glob patterns (e.g., "*timeout*", "*throttle*")
